+++ test/selector-beta/dependencies.spec.js
@@ -11,7 +11,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 
 const sinon = require("sinon");
 
-const { Provider, SelectorBeta, providers } = require("../../src/index");
+const { Provider, SelectorBeta, providers, catchDependency } = require("../../src/index");
 
 describe("Selector dependencies", () => {
   const DEPENDENCY_1_RESULT = "dependency-1-result";
@@ -359,7 +359,7 @@ describe("Selector dependencies", () => {
       expect(spies.dependency4Read.callCount).toEqual(1);
     });
 
-    it("should execute dependency more than once when called multiple times in query changes", async () => {
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
       selector.query(QUERY).read();
       selector.query(QUERY).read();
       await selector.query(QUERY).read();
@@ -447,10 +447,11 @@ describe("Selector dependencies", () => {
         DEPENDENCY_1_RESULT,
         DEPENDENCY_2_RESULT,
       ]);
-      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual(
-        [DEPENDENCY_1_RESULT, DEPENDENCY_2_RESULT],
-        [DEPENDENCY_3_RESULT]
-      );
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual([DEPENDENCY_3_RESULT]);
     });
 
     it("should receive dependencies results in last dependency", async () => {
@@ -473,7 +474,7 @@ describe("Selector dependencies", () => {
       expect(spies.dependency4Read.callCount).toEqual(1);
     });
 
-    it("should execute dependency more than once when called multiple times in query changes", async () => {
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
       selector.query(QUERY).read();
       selector.query(QUERY).read();
       await selector.query(QUERY).read();
@@ -511,11 +512,556 @@ describe("Selector dependencies", () => {
     });
   });
 
-  // TODO, test function returning catchDependencies
-  // TODO, test function returning array of functions
-  // TODO, test promise as dependency
-  // TODO, test value as dependency
-  // TODO, test function returning value
-  // TODO, test function returning promise
-  // TODO, test array of functions
+  describe("when defined as functions returning catchDependencies", () => {
+    const QUERY = { foo: "foo" };
+    let querySpy;
+    let querySpy2;
+    let querySpy3;
+    beforeEach(() => {
+      querySpy = sandbox.spy();
+      querySpy2 = sandbox.spy();
+      querySpy3 = sandbox.spy();
+      selector = new SelectorBeta(
+        (query) => {
+          querySpy(query);
+          return [
+            catchDependency(dependency1.query(query)),
+            catchDependency(dependency2.query(query)),
+          ];
+        },
+        (query, dependency1Result) => {
+          querySpy2({
+            query,
+            dependency1Result,
+          });
+          return catchDependency([dependency3.query(query)]);
+        },
+        (query, dependency1Result, dependency2Result) => {
+          querySpy3({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+          spies.dependency4Read();
+          return {
+            query,
+            dependency1Result,
+            dependency2Result,
+          };
+        }
+      );
+    });
+
+    it("should receive query as an argument in all dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy.getCall(0).args[0]).toEqual(QUERY);
+      expect(querySpy2.getCall(0).args[0].query).toEqual(QUERY);
+      expect(querySpy3.getCall(0).args[0].query).toEqual(QUERY);
+    });
+
+    it("should receive previous results as arguments in consequent dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual([DEPENDENCY_3_RESULT]);
+    });
+
+    it("should receive dependencies results in last dependency", async () => {
+      const result = await selector.read();
+      expect(result).toEqual({
+        query: {},
+        dependency1Result: [DEPENDENCY_1_RESULT, DEPENDENCY_2_RESULT],
+        dependency2Result: [DEPENDENCY_3_RESULT],
+      });
+    });
+
+    it("should execute dependency only once when called multiple times if query does not change", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(1);
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(1);
+    });
+
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      await selector.query({ foo: "foo2" }).read();
+      expect(spies.dependency1Read.callCount).toEqual(2);
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 1 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency1.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(2);
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 2 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency2.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(1);
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+  });
+
+  describe("when defined as functions returning an array of functions", () => {
+    const QUERY = { foo: "foo" };
+    let querySpy;
+    let querySpy2;
+    let querySpy3;
+    beforeEach(() => {
+      querySpy = sandbox.spy();
+      querySpy2 = sandbox.spy();
+      querySpy3 = sandbox.spy();
+      selector = new SelectorBeta(
+        () => {
+          return (query) => {
+            return [
+              (queryValue) => {
+                querySpy(queryValue);
+                return dependency1.query(queryValue);
+              },
+              dependency2.query(query),
+            ];
+          };
+        },
+        () => {
+          return (query, dependency1Result) => {
+            querySpy2({
+              query,
+              dependency1Result,
+            });
+            return dependency3.query(query);
+          };
+        },
+        (query, dependency1Result, dependency2Result) => {
+          querySpy3({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+          spies.dependency4Read();
+          return {
+            query,
+            dependency1Result,
+            dependency2Result,
+          };
+        }
+      );
+    });
+
+    it("should receive query as an argument in all dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy.getCall(0).args[0]).toEqual(QUERY);
+      expect(querySpy2.getCall(0).args[0].query).toEqual(QUERY);
+      expect(querySpy3.getCall(0).args[0].query).toEqual(QUERY);
+    });
+
+    it("should receive previous results as arguments in consequent dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual(DEPENDENCY_3_RESULT);
+    });
+
+    it("should receive dependencies results in last dependency", async () => {
+      const result = await selector.read();
+      expect(result).toEqual({
+        query: {},
+        dependency1Result: [DEPENDENCY_1_RESULT, DEPENDENCY_2_RESULT],
+        dependency2Result: DEPENDENCY_3_RESULT,
+      });
+    });
+
+    it("should execute dependency only once when called multiple times if query does not change", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(1);
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(1);
+    });
+
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      await selector.query({ foo: "foo2" }).read();
+      expect(spies.dependency1Read.callCount).toEqual(2);
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 1 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency1.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(2);
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 2 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency2.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency1Read.callCount).toEqual(1);
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+  });
+
+  describe("when defined as functions returning an array of data-provider instances and promises", () => {
+    const QUERY = { foo: "foo" };
+    let querySpy;
+    let querySpy2;
+    let querySpy3;
+    beforeEach(() => {
+      querySpy = sandbox.spy();
+      querySpy2 = sandbox.spy();
+      querySpy3 = sandbox.spy();
+      selector = new SelectorBeta(
+        (query) => {
+          querySpy(query);
+          return [
+            new Promise((resolve) => {
+              setTimeout(() => {
+                resolve(DEPENDENCY_1_RESULT);
+              }, 1000);
+            }),
+            dependency2.query(query),
+          ];
+        },
+        (query, dependency1Result) => {
+          querySpy2({
+            query,
+            dependency1Result,
+          });
+          return [dependency3.query(query)];
+        },
+        (query, dependency1Result, dependency2Result) => {
+          querySpy3({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+          spies.dependency4Read();
+          return {
+            query,
+            dependency1Result,
+            dependency2Result,
+          };
+        }
+      );
+    });
+
+    it("should receive query as an argument in all dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy.getCall(0).args[0]).toEqual(QUERY);
+      expect(querySpy2.getCall(0).args[0].query).toEqual(QUERY);
+      expect(querySpy3.getCall(0).args[0].query).toEqual(QUERY);
+    });
+
+    it("should receive previous results as arguments in consequent dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual([DEPENDENCY_3_RESULT]);
+    });
+
+    it("should receive dependencies results in last dependency", async () => {
+      const result = await selector.read();
+      expect(result).toEqual({
+        query: {},
+        dependency1Result: [DEPENDENCY_1_RESULT, DEPENDENCY_2_RESULT],
+        dependency2Result: [DEPENDENCY_3_RESULT],
+      });
+    });
+
+    it("should execute dependency only once when called multiple times if query does not change", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(1);
+    });
+
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      await selector.query({ foo: "foo2" }).read();
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 2 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency2.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+  });
+
+  describe("when defined as functions returning an array of data-provider instances and values", () => {
+    const QUERY = { foo: "foo" };
+    const DEPENDENCY_1_STATIC_RESULT = ["FOO", "FOO2"];
+    const DEPENDENCY_3_STATIC_RESULT = "VAR";
+    let querySpy;
+    let querySpy2;
+    let querySpy3;
+    beforeEach(() => {
+      querySpy = sandbox.spy();
+      querySpy2 = sandbox.spy();
+      querySpy3 = sandbox.spy();
+      selector = new SelectorBeta(
+        (query) => {
+          querySpy(query);
+          return [DEPENDENCY_1_STATIC_RESULT, dependency2.query(query)];
+        },
+        (query, dependency1Result) => {
+          querySpy2({
+            query,
+            dependency1Result,
+          });
+          return DEPENDENCY_3_STATIC_RESULT;
+        },
+        (query, dependency1Result, dependency2Result) => {
+          querySpy3({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+          spies.dependency4Read();
+          return {
+            query,
+            dependency1Result,
+            dependency2Result,
+          };
+        }
+      );
+    });
+
+    it("should receive query as an argument in all dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy.getCall(0).args[0]).toEqual(QUERY);
+      expect(querySpy2.getCall(0).args[0].query).toEqual(QUERY);
+      expect(querySpy3.getCall(0).args[0].query).toEqual(QUERY);
+    });
+
+    it("should receive previous results as arguments in consequent dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_STATIC_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_STATIC_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual(DEPENDENCY_3_STATIC_RESULT);
+    });
+
+    it("should receive dependencies results in last dependency", async () => {
+      const result = await selector.read();
+      expect(result).toEqual({
+        query: {},
+        dependency1Result: [DEPENDENCY_1_STATIC_RESULT, DEPENDENCY_2_RESULT],
+        dependency2Result: DEPENDENCY_3_STATIC_RESULT,
+      });
+    });
+
+    it("should execute dependency only once when called multiple times if query does not change", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(1);
+    });
+
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      await selector.query({ foo: "foo2" }).read();
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 2 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency2.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+  });
+
+  describe("when defined as a Promise resolving an array of data-provider instances and values", () => {
+    const QUERY = { foo: "foo" };
+    let querySpy2;
+    let querySpy3;
+    beforeEach(() => {
+      querySpy2 = sandbox.spy();
+      querySpy3 = sandbox.spy();
+      selector = new SelectorBeta(
+        new Promise((resolve) => {
+          setTimeout(() => {
+            resolve([DEPENDENCY_1_RESULT, dependency2]);
+          }, 1000);
+        }),
+        (query, dependency1Result) => {
+          return [
+            new Promise((resolve) => {
+              setTimeout(() => {
+                querySpy2({
+                  query,
+                  dependency1Result,
+                });
+                resolve(dependency3.query(query));
+              }, 1000);
+            }),
+          ];
+        },
+        (query, dependency1Result, dependency2Result) => {
+          querySpy3({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+          spies.dependency4Read();
+          return Promise.resolve({
+            query,
+            dependency1Result,
+            dependency2Result,
+          });
+        }
+      );
+    });
+
+    it("should receive query as an argument in all dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].query).toEqual(QUERY);
+      expect(querySpy3.getCall(0).args[0].query).toEqual(QUERY);
+    });
+
+    it("should receive previous results as arguments in consequent dependencies", async () => {
+      await selector.query(QUERY).read();
+      expect(querySpy2.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency1Result).toEqual([
+        DEPENDENCY_1_RESULT,
+        DEPENDENCY_2_RESULT,
+      ]);
+      expect(querySpy3.getCall(0).args[0].dependency2Result).toEqual([DEPENDENCY_3_RESULT]);
+    });
+
+    it("should receive dependencies results in last dependency", async () => {
+      const result = await selector.read();
+      expect(result).toEqual({
+        query: {},
+        dependency1Result: [DEPENDENCY_1_RESULT, DEPENDENCY_2_RESULT],
+        dependency2Result: [DEPENDENCY_3_RESULT],
+      });
+    });
+
+    it("should execute dependency only once when called multiple times if query does not change", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(1);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(1);
+    });
+
+    it("should execute dependency more than once when called multiple times if query changes", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      await selector.query({ foo: "foo2" }).read();
+      expect(spies.dependency3Read.callCount).toEqual(2);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+
+    it("should clean cache when dependency 2 cache is clean", async () => {
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      selector.query(QUERY).read();
+      await selector.query(QUERY).read();
+      dependency2.cleanCache();
+      await selector.query(QUERY).read();
+      expect(spies.dependency2Read.callCount).toEqual(2);
+      expect(spies.dependency3Read.callCount).toEqual(1);
+      expect(spies.dependency4Read.callCount).toEqual(2);
+    });
+  });
+
+  // TODO, test functions returning value
+  // TODO, test Promise cache (cleanCache) should clean it
+  // TODO, test promise catched
+  // TODO, test promise catched returning error
 });